GRAPHICS: Let Font take uint32 as character codes.
This is required to support UTF-32 strings but does not make them work
automatically!

+++ graphics/font.cpp
@@ -26,7 +26,7 @@
 
 namespace Graphics {
 
-int Font::getKerningOffset(byte left, byte right) const {
+int Font::getKerningOffset(uint32 left, uint32 right) const {
 	return 0;
 }
